+++ src/lib/ui/icon.ts
@@ -1,7 +1,7 @@
 import type { HComponentFn } from "@/lib/core/component";
-import { element } from "@/lib/core/component";
+import { component, element } from "@/lib/core/component";
 import { Link } from "@/lib/core/elements";
-import { component, registerComponent } from "@/lib/core/store";
+import { registerComponent } from "@/lib/core/store";
 import type { Store } from "@/lib/core/store";
 
 export type HIconType = "brands" | "solid";
